+++ src/components/common/BtnTask/BtnTask.tsx
@@ -4,12 +4,10 @@ import React, { useState } from 'react';
 
 import Modal from '../modal/Modal';
 
+import IconHoverContainer from './IconHoverContainer';
+
 import Icons from '@/assets/svg/index';
 import BtnDate from '@/components/common/BtnDate/BtnDate';
-import StatusDoneBtn from '@/components/common/button/statusBtn/StatusDoneBtn';
-import StatusInProgressBtn from '@/components/common/button/statusBtn/StatusInProgressBtn';
-import StatusStagingBtn from '@/components/common/button/statusBtn/StatusStagingBtn';
-import StatusTodoBtn from '@/components/common/button/statusBtn/StatusTodoBtn';
 import { TaskType } from '@/types/tasks/taskType';
 
 interface BtnTaskProps extends TaskType {
@@ -19,7 +17,6 @@ interface BtnTaskProps extends TaskType {
 interface BorderColorProps {
 	isHovered: boolean;
 	isClicked: boolean;
-	iconHovered: boolean;
 	theme: Theme;
 	btnType: string;
 }
@@ -29,8 +26,6 @@ function BtnTask(props: BtnTaskProps) {
 	const [isModalOpen, setModalOpen] = useState(false);
 	const [isClicked, setIsClicked] = useState(false);
 	const [isHovered, setIsHovered] = useState(false);
-	const [iconHovered, setIconHovered] = useState(false);
-	const [iconClicked, setIconClicked] = useState(false);
 
 	const [top, setTop] = useState(0);
 	const [left, setLeft] = useState(0);
@@ -65,60 +60,12 @@ function BtnTask(props: BtnTaskProps) {
 	const closeModal = () => {
 		setModalOpen(false);
 	};
-	const stopPropagation = (e: React.MouseEvent) => {
-		e.stopPropagation();
-	};
-
-	const handleIconMouseEnter = () => {
-		setIconHovered(true);
-	};
-
-	const handleIconMouseLeave = () => {
-		setIconHovered(false);
-	};
-
-	const handleIconClick = (e: React.MouseEvent) => {
-		if (iconClicked) {
-			setIconClicked(false);
-		} else {
-			setIconClicked(true);
-		}
-		stopPropagation(e);
-	};
-
-	const renderStatusButton = () => {
-		if (btnType === 'delayed') {
-			return <StagingIconHoverIndicator />;
-		}
-
-		if (!iconHovered && !iconClicked) {
-			return <IconHoverIndicator />;
-		}
-
-		if (btnType === 'staging') {
-			return <StatusStagingBtn />;
-		}
-
-		if (btnType === 'target') {
-			switch (status) {
-				case '완료':
-					return <StatusDoneBtn />;
-				case '진행중':
-					return <StatusInProgressBtn />;
-				default:
-					return <StatusTodoBtn />;
-			}
-		}
-
-		return null;
-	};
 
 	return (
 		<ModalLayout>
 			<BtnTaskLayout
 				isClicked={isClicked}
 				isHovered={isHovered}
-				iconHovered={iconHovered}
 				btnType={btnType}
 				onDoubleClick={handleDoubleClick}
 				onClick={handleClick}
@@ -135,13 +82,7 @@ function BtnTask(props: BtnTaskProps) {
 						isDelayed={btnType === 'delayed'}
 					/>
 				</BtnTaskContainer>
-				<IconHoverContainer
-					onClick={handleIconClick}
-					onMouseEnter={handleIconMouseEnter}
-					onMouseLeave={handleIconMouseLeave}
-				>
-					{renderStatusButton()}
-				</IconHoverContainer>
+				<IconHoverContainer btnType={btnType} status={status} />
 			</BtnTaskLayout>
 			<Modal isOpen={isModalOpen} sizeType={{ type: 'short' }} top={top} left={left} onClose={closeModal} />
 		</ModalLayout>
@@ -150,7 +91,7 @@ function BtnTask(props: BtnTaskProps) {
 
 export default BtnTask;
 
-const getBorderColor = ({ isHovered, isClicked, iconHovered, theme, btnType }: BorderColorProps) => {
+const getBorderColor = ({ isHovered, isClicked, theme, btnType }: BorderColorProps) => {
 	const defaultColor = theme.palette.Grey.Grey1;
 	const hoverColor = theme.palette.Blue.Blue3;
 	const clickColor = theme.palette.Primary;
@@ -160,7 +101,7 @@ const getBorderColor = ({ isHovered, isClicked, iconHovered, theme, btnType }: B
 		borderColor = orangeColor;
 	} else if (isClicked) {
 		borderColor = clickColor;
-	} else if (iconHovered || isHovered) {
+	} else if (isHovered) {
 		borderColor = hoverColor;
 	}
 	return css`
@@ -175,7 +116,6 @@ const ModalLayout = styled.div`
 const BtnTaskLayout = styled('div', { target: 'BtnTaskLayout' })<{
 	isClicked: boolean;
 	isHovered: boolean;
-	iconHovered: boolean;
 	btnType: string;
 }>`
 	display: flex;
@@ -207,35 +147,6 @@ const BtnTaskTextWrapper = styled.div<{ isDescription: boolean }>`
 	padding-left: ${({ isDescription }) => (isDescription ? '0rem' : '0.4rem')};
 
 	${({ theme }) => theme.fontTheme.LABEL_03};
-`;
-
-const IconHoverContainer = styled('div', { target: 'IconHoverContainer' })`
-	display: flex;
-	gap: 0.4rem;
-	align-items: center;
-	padding: 0 1.2rem 0 0.8rem;
-`;
-
-const IconHoverCss = css`
-	display: flex;
-	gap: 0.6rem;
-	align-items: center;
-	justify-content: center;
-	width: 1.4rem;
-	height: 1.4rem;
-	padding: 0.3rem;
-`;
-
-const IconHoverIndicator = styled(Icons.Icn_hover_indicator, { target: 'IconHoverIndicator' })`
-	${IconHoverCss}
-`;
-
-const StagingIconHoverIndicator = styled(Icons.Icn_hover_indicator)`
-	${IconHoverCss}
-
-	circle {
-		stroke: ${({ theme }) => theme.palette.Orange.Orange8};
-	}
 `;
 
 const IconFile = styled(Icons.IcnFile)`
